create display pane and begin to show rooms
