+++ perikaryon-view/package.json
@@ -11,6 +11,7 @@
     "popper.js": "^1.15.0",
     "react": "^16.9.0",
     "react-beautiful-dnd": "^11.0.5",
+    "react-bootstrap": "^1.0.0-beta.12",
     "react-dom": "^16.9.0",
     "react-draggable": "^3.3.2",
     "react-grid-layout": "^0.16.6",
