+++ perikaryon-view/public/index.html
@@ -25,7 +25,8 @@
     -->
     <title>React App</title>
   </head>
-  <body id="root">
+  <body>
+    <div id="root"></div>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     
     <!--
