+++ perikaryon-view/src/App.scss
@@ -1,7 +1,7 @@
 @import "node_modules/bootstrap/scss/bootstrap";
 @import "../node_modules/react-grid-layout/css/styles.css";
 @import "../node_modules/react-resizable/css/styles.css";
-@import '../node_modules/jsoneditor-react/es/editor.min.css';
+@import "../node_modules/jsoneditor-react/es/editor.min.css";
 
 $ranvier-light-green: rgba(123, 187, 117, 1);
 $ranvier-green: #68a063;
@@ -30,126 +30,128 @@ $ranvier-selected-orange: #ff7d04;
 }
 div.room {
   cursor: pointer;
-  padding:4px;
+  padding: 4px;
   text-align: center;
   background-color: $ranvier-green;
   border: 3px solid $ranvier-dark-green;
   border-radius: 15px;
-  
 }
-html, body, #root, .App {
-  width:100%
+html,
+body,
+#root,
+.App {
+  width: 100%;
 }
-.container{
-  width:100%;
+.container {
+  width: 100%;
 }
-#itemDiv:before{
-  content:"Items";
+#itemDiv:before {
+  content: "Items";
   text-align: center;
   background-color: $ranvier-green;
   border: 3px solid $ranvier-dark-green;
   display: block;
 }
-#itemDiv.col-xl{
+#itemDiv.col-xl {
   padding-left: 0px;
-  padding-right:0px;
+  padding-right: 0px;
 }
 #descriptionDiv:before {
-  content:"Description";
+  content: "Description";
   text-align: center;
   background-color: $ranvier-green;
   border: 3px solid $ranvier-dark-green;
   display: block;
 }
-#descriptionDiv{
+#descriptionDiv {
   padding-left: 0px;
-  padding-right:0px;
+  padding-right: 0px;
   width: 40%;
 }
-#npcDiv:before{
-  content:"NPCs";
+#npcDiv:before {
+  content: "NPCs";
   text-align: center;
   background-color: $ranvier-green;
   border: 3px solid $ranvier-dark-green;
   display: block;
 }
-#npcDiv{
+#npcDiv {
   padding-left: 0px;
-  padding-right:0px;
-  width:40%;
+  padding-right: 0px;
+  width: 40%;
 }
-.dashbutton{
+.dashbutton {
   float: right;
 }
-#dashboard{
-  position:fixed;
-  bottom:0;
-  width:100%;
+#dashboard {
+  position: fixed;
+  bottom: 0;
+  width: 100%;
 }
-#reactgrid{
+#reactgrid {
   margin-left: auto;
   margin-right: auto;
 }
-#reactgrid::after{
+#reactgrid::after {
   clear: both;
 }
-#topdash{
+#topdash {
   margin-left: auto;
   margin-right: auto;
   text-align: center;
 }
-.collapse.show{
+.collapse.show {
   visibility: visible;
 }
-.topdashbtn{
-  margin:3px;
+.topdashbtn {
+  margin: 3px;
 }
-.topdashbtn.active{
+.topdashbtn.active {
   background-color: $ranvier-selected-orange;
 }
-select:invalid{
+select:invalid {
   color: gray;
 }
-div.room.Selected{
+div.room.Selected {
   background-color: $ranvier-selected-orange;
   border: 3px solid $ranvier-selected-orange;
-  box-shadow: 2px 2px 2px rgba(0,0,0,0.5);
+  box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.5);
 }
-div.room.Selected:hover{
+div.room.Selected:hover {
   background-color: $ranvier-selected-orange;
 }
-div.room:hover{
+div.room:hover {
   background-color: $ranvier-light-green;
-  box-shadow: 2px 2px 2px rgba(0,0,0,0.5);
+  box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.5);
 }
-#areaDropdown{
-  width:200px;
+#areaDropdown {
+  width: 200px;
 }
-#floorDropdown{
-  width:100px;
+#floorDropdown {
+  width: 100px;
 }
-#title{
-  width:60%;
-  font-size:3vw;
-  margin:0;
+#title {
+  width: 60%;
+  font-size: 3vw;
+  margin: 0;
 }
-#topNavBar{
+#topNavBar {
   padding-top: 0px;
   padding-bottom: 0px;
 }
-#roomButtons{
-  padding-right:30px;
+#roomButtons {
+  padding-right: 30px;
 }
-#buttondiv{
-  margin:0;
-  padding:0;
+#buttondiv {
+  margin: 0;
+  padding: 0;
 }
-#roomDescription{
-  width:100%;
-  height:100%;
-  resize:vertical;
-  -webkit-box-sizing:border-box;
-  -moz-box-sizing:border-box;
+#roomDescription {
+  width: 100%;
+  height: 100%;
+  resize: vertical;
+  -webkit-box-sizing: border-box;
+  -moz-box-sizing: border-box;
   box-sizing: border-box;
 }
 .App-link {
@@ -163,4 +165,11 @@ div.room:hover{
   to {
     transform: rotate(360deg);
   }
+}
+
+.header {
+  text-align: center;
+  background-color: $ranvier-green;
+  border: 3px solid $ranvier-dark-green;
+  display: block;
 }
