+++ perikaryon-view/src/components/Header.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <div className='header'>
+      <h1>Ranvier Mapping</h1>
+    </div>
+  );
+}
+
+export default Header;
